Mild cleanup.

+++ inc/cdam_const_config.h
@@ -9,15 +9,15 @@ namespace cdam
 // Seconds for timeout of Published Spark Events
 const uint8_t kServerTTLDefault = 60;
 // How often to update the Printer state.
-const unsigned int kIntervalPrinterMillis = 50;
+const uint16_t kIntervalPrinterMillis = 50;
 // How often to update the Keypad state.
-const unsigned int kIntervalKeypadMillis = 5;
+const uint16_t kIntervalKeypadMillis = 5;
 // How often to update the Coin Acceptor state.
-const unsigned int kIntervalCoinAcceptorMillis = 5;
+const uint16_t kIntervalCoinAcceptorMillis = 5;
 // How often to allow printing Insert Coin message.
-const unsigned int kIntervalCoinInsertMessage = 10000;
+const uint16_t kIntervalCoinInsertMessage = 10000;
 // Delay until printing out 'press any button' message.
-const unsigned int kIntervalPressAnyButton = 30000;
+const uint16_t kIntervalPressAnyButton = 30000;
 // Min value for coin sensing.
 const uint8_t kCoinAcceptorSenseMin = 15;
 // Max value for coin sensing.
